wifi: ieee80211: split HE definitions out
[ Upstream commit 02a2cf302557eb59794bba0b05d6755f44928d78 ]

The ieee80211.h file has gotten very long, continue splitting
it by putting HE definitions into a separate file.

Link: https://patch.msgid.link/20251105153843.6998c0802104.I3dd7cfea6abbd118b999ecdedd48437d39cb0533@changeid
Stable-dep-of: cb0caadb64ca ("wifi: ieee80211: fix definition of EHT-MCS 15 in MRU")
(cherry picked from commit df5720d35848be92944767a6d5ae20ec9ed347ce)
